Fix doctor registration issues and refactor into components
- Fixed TypeScript errors in DoctorForm by properly typing state callbacks
- Refactored monolithic DoctorForm into separate step components:
  - DoctorStep1.tsx (Personal Information)
  - DoctorStep2.tsx (Professional Details)
  - DoctorStep3.tsx (Practice Information)
- Maintained all validation logic and form functionality
- Added proper type safety throughout all components
- Fixed consultationFee typo in validation
- Preserved all existing styling and layout
